Add new method for end-of-level, remove broken
This has been tested and appears working by watching the fanfare queue byte at `0x4E4`, so it will correctly split almost all normal stage and fortress exits by watching for the appropriate music to begin playing. AFAICT, this is frame-perfect.

It will not trigger on 1-3 whistle or any of the ships or hand spaces on 8 (as they don't play music), but this is a significant improvement over the current version. Still researching those.